chore - extract the remaining Body IR helpers into their own modules (#1101)
The construct split left `body_ir.rs` at 2,541 lines: the builder and its
scaffolding, but also every free helper the lowering shares. Those are now
modules too -- `collect`, `bodies`, `args`, `reads`, `free_vars`, `refusals`,
and `primitives` -- and the comprehension terminal and default-statement
helpers move beside the construct modules that own them.

`body_ir.rs` is 876 lines: the module entry point, the lowering-facts bundle,
the `BodyBuilder` struct, and its scaffolding (scopes, locals, temps, ownership
facts, scope drops). Everything else is one module per concern.

Visibility follows the tree rather than being widened: helpers are `pub(super)`,
and each module imports the helper modules it actually uses. The parent
deliberately does not re-export them -- a re-export at `pub(super)` reexports
nothing, because the items are only visible within `body_ir` in the first place,
which strict clippy correctly rejects.
