+++ src/frontend/body_ir/async_.rs
@@ -1,5 +1,6 @@
 //! Lowering for `await` and `race` suspension points.
 
+use super::reads::*;
 use super::*;
 
 impl<'type_info, 'source> BodyBuilder<'type_info, 'source> {
