+++ src/frontend/body_ir/control_flow.rs
@@ -1,5 +1,8 @@
 //! Lowering for conditional and looping control flow, and the iteration protocol behind `for`.
 
+use super::args::*;
+use super::reads::*;
+use super::refusals::*;
 use super::*;
 
 impl<'type_info, 'source> BodyBuilder<'type_info, 'source> {
