+++ src/frontend/body_ir/literals.rs
@@ -1,5 +1,8 @@
 //! Lowering for literal and aggregate forms: lists, dicts, f-strings, slices, `?`, constructors, surface exprs.
 
+use super::args::*;
+use super::primitives::*;
+use super::refusals::*;
 use super::*;
 
 impl<'type_info, 'source> BodyBuilder<'type_info, 'source> {
